feat: add recursive creation of files and folders in 04 06

+++ 04-copy-directory/index.js
@@ -4,7 +4,7 @@ const path = require('path');
 const fullFolderPath = path.join(__dirname, 'files');
 const newFolderPath = path.join(__dirname, 'files-copy');
 
-const newFolderCreator = () => {
+const newFolderCreator = (fullFolderPath, newFolderPath) => {
   fs.mkdir(newFolderPath, { recursive: true }, err => {
     if (err) throw new Error('Folder is created');
 
@@ -12,10 +12,24 @@ const newFolderCreator = () => {
       if (errFiles) throw new Error(`Error with reading files in ${newFolderPath}`);
 
       files.forEach(file => {
-        fs.access(path.join(fullFolderPath, file), err => {
-          if (err) {
-            fs.unlink(path.join(newFolderPath, file), deletingErr => {
-              if (deletingErr) throw new Error(`Error with deleting file: ${file}`);
+        fs.stat(path.join(newFolderPath, file), (errInfo, stats) => {
+          if (errInfo) throw new Error(`Error with ${stats}`);
+
+          if (stats.isFile()) {
+            fs.access(path.join(fullFolderPath, file), err => {
+              if (err) {
+                fs.unlink(path.join(newFolderPath, file), deletingErr => {
+                  if (deletingErr) throw new Error(`Error with deleting file: ${file}`);
+                });
+              }
+            });
+          } else if (stats.isDirectory()) {
+            fs.access(path.join(fullFolderPath, file), err => {
+              if (err) {
+                fs.rm(path.join(newFolderPath, file), { recursive: true }, deletingErr => {
+                  if (deletingErr) throw new Error(`Error with deleting file: ${file}`);
+                });
+              }
             });
           }
         });
@@ -24,20 +38,29 @@ const newFolderCreator = () => {
   });
 };
 
-const readWriteStream = () => {
+const readWriteStream = (fullFolderPath, newFolderPath) => {
   fs.readdir(fullFolderPath, (err, files) => {
     if (err) throw new Error('Error with file');
 
     files.forEach(file => {
-      const readStream = fs.createReadStream(path.join(fullFolderPath, file));
-      const writeStream = fs.createWriteStream(path.join(newFolderPath, file));
-      readStream.pipe(writeStream);
+      fs.stat(path.join(fullFolderPath, file), (errInfo, stats) => {
+        if (errInfo) throw new Error(`Error with ${stats}`);
+
+        if (stats.isFile()) {
+          const readStream = fs.createReadStream(path.join(fullFolderPath, file));
+          const writeStream = fs.createWriteStream(path.join(newFolderPath, file));
+          readStream.pipe(writeStream);
+        } else if (stats.isDirectory()) {
+          newFolderCreator(path.join(fullFolderPath, file), path.join(newFolderPath, file));
+          readWriteStream(path.join(fullFolderPath, file), path.join(newFolderPath, file));
+        }
+      });
     });
   });
 };
 
 const copyFiles = async () => {
-  await newFolderCreator();
-  await readWriteStream();
+  await newFolderCreator(fullFolderPath, newFolderPath);
+  await readWriteStream(fullFolderPath, newFolderPath);
 };
 copyFiles();
